+++ Robust.Shared/Physics/Systems/SharedBroadphaseSystem.cs
@@ -165,10 +165,9 @@ namespace Robust.Shared.Physics.Systems
 
             // FindNewContacts is inherently going to be a lot slower than Box2D's normal version so we need
             // to cache a bunch of stuff to make up for it.
-            var contactManager = component.ContactManager;
 
             // Handle grids first as they're not stored on map broadphase at all.
-            HandleGridCollisions(mapId, contactManager, movedGrids, physicsQuery, xformQuery);
+            HandleGridCollisions(mapId, movedGrids, physicsQuery, xformQuery);
 
             // EZ
             if (moveBuffer.Count == 0)
@@ -249,7 +248,7 @@ namespace Robust.Shared.Physics.Systems
                         _physicsSystem.WakeBody(otherBody, force: true);
                     }
 
-                    contactManager.AddPair(proxyA, other);
+                    _physicsSystem.AddPair(proxyA, other);
                 }
 
                 _bufferPool.Return(contactBuffer[i]);
@@ -264,7 +263,6 @@ namespace Robust.Shared.Physics.Systems
 
         private void HandleGridCollisions(
             MapId mapId,
-            ContactManager contactManager,
             HashSet<MapGridComponent> movedGrids,
             EntityQuery<PhysicsComponent> bodyQuery,
             EntityQuery<TransformComponent> xformQuery)
@@ -326,7 +324,7 @@ namespace Robust.Shared.Physics.Systems
                                             var otherAABB = otherFixture.Shape.ComputeAABB(otherTransform, j);
 
                                             if (!fixAABB.Intersects(otherAABB)) continue;
-                                            contactManager.AddPair(fixture, i, otherFixture, j, ContactFlags.Grid);
+                                            _physicsSystem.AddPair(fixture, i, otherFixture, j, ContactFlags.Grid);
                                             break;
                                         }
                                     }
@@ -400,7 +398,7 @@ namespace Robust.Shared.Physics.Systems
                 // Logger.DebugS("physics", $"Checking {proxy.Fixture.Body.Owner} against {other.Fixture.Body.Owner} at {aabb}");
 
                 if (tuple.proxy == other ||
-                    !ContactManager.ShouldCollide(tuple.proxy.Fixture, other.Fixture) ||
+                    !SharedPhysicsSystem.ShouldCollide(tuple.proxy.Fixture, other.Fixture) ||
                     tuple.proxy.Fixture.Body == other.Fixture.Body)
                 {
                     return true;
