+++ Robust.UnitTesting/Shared/Physics/Collision_Test.cs
@@ -22,11 +22,15 @@
 
 using System;
 using NUnit.Framework;
+using Robust.Shared.GameObjects;
+using Robust.Shared.Map;
 using Robust.Shared.Maths;
 using Robust.Shared.Physics;
 using Robust.Shared.Physics.Collision.Shapes;
+using Robust.Shared.Physics.Components;
 using Robust.Shared.Physics.Dynamics;
 using Robust.Shared.Physics.Systems;
+using Robust.UnitTesting.Server;
 
 namespace Robust.UnitTesting.Shared.Physics;
 
@@ -83,5 +87,48 @@ public sealed class Collision_Test
         Assert.That(MathF.Abs(massData2.Center.Y - center.Y), Is.LessThan(absTol + relTol * Math.Abs(center.Y)));
         Assert.That(MathF.Abs(massData2.Mass - mass), Is.LessThan(20.0f * (absTol + relTol * mass)));
         Assert.That(MathF.Abs(massData2.I - inertia), Is.LessThan(40.0f * (absTol + relTol * inertia)));
+    }
+
+    /// <summary>
+    /// Asserts that cross-map contacts correctly destroy
+    /// </summary>
+    [Test]
+    public void CrossMapContacts()
+    {
+        var sim = RobustServerSimulation.NewSimulation().InitializeInstance();
+        var entManager = sim.Resolve<IEntityManager>();
+        var mapManager = sim.Resolve<IMapManager>();
+        var fixtures = entManager.System<FixtureSystem>();
+        var physics = entManager.System<SharedPhysicsSystem>();
+        var xformSystem = entManager.System<SharedTransformSystem>();
+        var mapId = mapManager.CreateMap();
+        var mapId2 = mapManager.CreateMap();
+
+        var ent1 = entManager.SpawnEntity(null, new MapCoordinates(Vector2.Zero, mapId));
+        var ent2 = entManager.SpawnEntity(null, new MapCoordinates(Vector2.Zero, mapId));
+
+        var body1 = entManager.AddComponent<PhysicsComponent>(ent1);
+        physics.SetBodyType(body1, BodyType.Dynamic);
+        var body2 = entManager.AddComponent<PhysicsComponent>(ent2);
+        physics.SetBodyType(body2, BodyType.Dynamic);
+
+        fixtures.CreateFixture(body1, new Fixture(new PhysShapeCircle() { Radius = 1f }, 1, 0, true));
+        fixtures.CreateFixture(body2, new Fixture(new PhysShapeCircle() { Radius = 1f }, 0, 1, true));
+
+        physics.WakeBody(body1);
+        physics.WakeBody(body2);
+
+        Assert.That(body1.Awake && body2.Awake);
+        Assert.That(body1.ContactCount == 0 && body2.ContactCount == 0);
+
+        physics.Update(0.01f);
+
+        Assert.That(body1.ContactCount == 1 && body2.ContactCount == 1);
+
+        // Reparent body2 and assert the contact is destroyed
+        xformSystem.SetParent(ent2, mapManager.GetMapEntityId(mapId2));
+        physics.Update(0.01f);
+
+        Assert.That(body1.ContactCount == 0 && body2.ContactCount == 0);
     }
 }
